Create addDateSuffix.js
This script is designed to be used within a custom-coded workflow action (Operations Hub Professional) and adds the appropriate suffix depending on the day of the month. It relies on you already having the date stored in a single-line text property DD/MM/YYYY. Worth noting this assumes you are storing the date in DD/MM/YYYY format.