Create custom_object_assoc_workflow
With this custom code action you can run a workflow on the associations of any custom object. The code snippet will look at the custom object enrolled in the workflow, pull in any of its associations and push those over to another workflow of yours. For instance, a workflow for email marketing, or that will send out a notification to their owner, or that will update a property. Created by: Sam G, Skupos